Add files via upload
These files are the necessary input files for the notebooks in this repository. PctUrbanRural_County-2.csv and lead-tool-map-data.csv are input files for ACS_Notebook.ipynb, ACS_data.zip is the outputted file from ACS_Notebook.ipynb and is used as an input for Energy_Poverty_California-Notebook.ipynb, as well as recs2015_public_v4.csv.